+++ XmlSpawner/XmlSpawner Core/XmlMobiles/TalkingBaseVendor.cs
@@ -757,7 +757,7 @@ public abstract class TalkingBaseVendor : BaseVendor
     {
         base.Serialize( writer );
 
-        writer.Write( (int) 7 ); // version
+        writer.Write( 7 ); // version
 
         // version 7
         writer.Write( m_EItemID);
