+++ XmlSpawner/XmlSpawner Core/XmlMobiles/XmlQuestNPC.cs
@@ -277,7 +277,7 @@ public class XmlQuestNPC : TalkingBaseCreature
     {
         base.Serialize( writer );
 
-        writer.Write( (int) 0 ); // version
+        writer.Write( 0 ); // version
 
     }
